// src/api/generated/models/myAppCaseInfo.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AppAuthor } from './appAuthor';

export interface MyAppCaseInfo {
  id?: string;
  appId?: string;
  title?: string;
  summary?: string;
  featured?: boolean;
  appName?: string;
  previewUrl?: string;
  coverUrl?: string;
  author?: AppAuthor;
  reviewedAt?: string;
  createdAt?: string;
  status?: string;
  reviewRemark?: string;
  updatedAt?: string;
}