// src/api/generated/models/pageResultMyAppCaseInfo.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { MyAppCaseInfo } from './myAppCaseInfo';

export interface PageResultMyAppCaseInfo {
  list?: MyAppCaseInfo[];
  total?: string;
  pageNum?: number;
  pageSize?: number;
  totalPages?: number;
  hasNext?: boolean;
  hasPrevious?: boolean;
}